+++ .idea/modules/app/VitalisApp.app.iml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<module version="4">
+  <component name="ModuleClassLoaderOverlays">
+    <paths>
+      <option value="C:\Users\Lenovo\AppData\Local\Temp\overlay8765218706897119808" />
+    </paths>
+  </component>
+</module>
